feat: implement search and intelligence handlers for MCP server
- Introduced SearchHandler and IntelligenceHandler classes to manage search-related and cross-document intelligence operations, respectively.
- Refactored MCPHandler to utilize the new handlers for improved modularity and maintainability.
- Added comprehensive response formatting through MCPFormatters for consistent output across search and intelligence functionalities.
- Implemented structured result creation for compliance with MCP protocol standards.
- Enhanced error handling and logging for better debugging and reliability.
- Added unit tests to validate the functionality of the new handlers and ensure backward compatibility.
